# openapi.yaml
openapi: 3.1.0
info:
  title: Schedule RKSI API
  version: 0.1.0
paths:
...
  "/schedule/public/by-group":
    get:
      tags:
      - schedule
      summary: Get Public By Group
      operationId: get_public_by_group_schedule_public_by_group_get
      parameters:
      - name: group_id
        in: query
        required: true
        schema:
          type: integer
          title: Group Id
      - name: week
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PublicScheduleWeekResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/schedule/public/by-teacher":
    get:
      tags:
      - schedule
      summary: Get Public By Teacher
      operationId: get_public_by_teacher_schedule_public_by_teacher_get
      parameters:
      - name: teacher_id
        in: query
        required: true
        schema:
          type: integer
          title: Teacher Id
      - name: week
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PublicScheduleWeekResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/schedule/public/by-room":
    get:
      tags:
      - schedule
      summary: Get Public By Room
      operationId: get_public_by_room_schedule_public_by_room_get
      parameters:
      - name: room_id
        in: query
        required: true
        schema:
          type: integer
          title: Room Id
      - name: week
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PublicScheduleWeekResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/schedule/lessons":
    get:
      tags:
      - schedule
      summary: Get Lessons
      operationId: get_lessons_schedule_lessons_get
      parameters:
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Date
      - name: time_slot
        in: query
        required: true
        schema:
          type: integer
          title: Time Slot
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ScheduleSlotRoomResponse"
                title: Response Get Lessons Schedule Lessons Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    post:
      tags:
      - schedule
      summary: Create Lesson
      operationId: create_lesson_schedule_lessons_post
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LessonCreateRequest"
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LessonMutationResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/schedule/problems":
    get:
      tags:
      - schedule
      summary: Get Problems
      operationId: get_problems_schedule_problems_get
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ScheduleProblemResponse"
                title: Response Get Problems Schedule Problems Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/schedule/lessons/{lesson_id}":
    patch:
      tags:
      - schedule
      summary: Update Lesson
      operationId: update_lesson_schedule_lessons__lesson_id__patch
      parameters:
      - name: lesson_id
        in: path
        required: true
        schema:
          type: integer
          title: Lesson Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LessonUpdateRequest"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LessonMutationResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
      - schedule
      summary: Delete Lesson
      operationId: delete_lesson_schedule_lessons__lesson_id__delete
      parameters:
      - name: lesson_id
        in: path
        required: true
        schema:
          type: integer
          title: Lesson Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LessonCreateRequest:
      properties:
        group_name:
          type: string
          title: Group Name
        course:
          type: integer
          title: Course
          default: 0
        faculty:
          type: string
          title: Faculty
          default: ''
        subject:
          type: string
          title: Subject
        teacher_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Name
        teacher_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Id
        teacher_post:
          type: string
          title: Teacher Post
          default: ''
        room_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Name
        date:
          type: string
          format: date
          title: Date
        time_start:
          type: string
          format: time
          title: Time Start
        time_end:
          type: string
          format: time
          title: Time End
        weekday:
          type: integer
          maximum: 7.0
          minimum: 1.0
          title: Weekday
        week_number:
          type: integer
          title: Week Number
        time_slot:
          type: integer
          minimum: 1.0
          title: Time Slot
        subgroup:
          type: integer
          minimum: 0.0
          title: Subgroup
          default: 0
        lesson_type:
          type: string
          title: Lesson Type
          default: ''
      type: object
      required:
      - group_name
      - subject
      - date
      - time_start
      - time_end
      - weekday
      - week_number
      - time_slot
      title: LessonCreateRequest
    LessonMutationResponse:
      properties:
        id:
          type: integer
          title: Id
        group_name:
          type: string
          title: Group Name
        subject:
          type: string
          title: Subject
        teacher_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Name
        teacher_is_absent:
          type: boolean
          title: Teacher Is Absent
          default: false
        teacher_absence_reason:
          type: string
          title: Teacher Absence Reason
          default: ''
        room_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Name
        date:
          type: string
          format: date
          title: Date
        time_start:
          type: string
          format: time
          title: Time Start
        time_end:
          type: string
          format: time
          title: Time End
        weekday:
          type: integer
          title: Weekday
        week_number:
          type: integer
          title: Week Number
        time_slot:
          type: integer
          title: Time Slot
        subgroup:
          type: integer
          title: Subgroup
        lesson_type:
          type: string
          title: Lesson Type
        warnings:
          items:
            "$ref": "#/components/schemas/ScheduleProblemResponse"
          type: array
          title: Warnings
      type: object
      required:
      - id
      - group_name
      - subject
      - teacher_name
      - room_name
      - date
      - time_start
      - time_end
      - weekday
      - week_number
      - time_slot
      - subgroup
      - lesson_type
      title: LessonMutationResponse
    LessonResponse:
      properties:
        id:
          type: integer
          title: Id
        group_name:
          type: string
          title: Group Name
        subject:
          type: string
          title: Subject
        teacher_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Name
        teacher_is_absent:
          type: boolean
          title: Teacher Is Absent
          default: false
        teacher_absence_reason:
          type: string
          title: Teacher Absence Reason
          default: ''
        room_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Name
        date:
          type: string
          format: date
          title: Date
        time_start:
          type: string
          format: time
          title: Time Start
        time_end:
          type: string
          format: time
          title: Time End
        weekday:
          type: integer
          title: Weekday
        week_number:
          type: integer
          title: Week Number
        time_slot:
          type: integer
          title: Time Slot
        subgroup:
          type: integer
          title: Subgroup
        lesson_type:
          type: string
          title: Lesson Type
      type: object
      required:
      - id
      - group_name
      - subject
      - teacher_name
      - room_name
      - date
      - time_start
      - time_end
      - weekday
      - week_number
      - time_slot
      - subgroup
      - lesson_type
      title: LessonResponse
    LessonUpdateRequest:
      properties:
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        course:
          anyOf:
          - type: integer
          - type: 'null'
          title: Course
        faculty:
          anyOf:
          - type: string
          - type: 'null'
          title: Faculty
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        teacher_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Name
        teacher_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Id
        teacher_post:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Post
        room_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Name
        date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
        time_start:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Time Start
        time_end:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Time End
        weekday:
          anyOf:
          - type: integer
            maximum: 7.0
            minimum: 1.0
          - type: 'null'
          title: Weekday
        week_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week Number
        time_slot:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Time Slot
        subgroup:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Subgroup
        lesson_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Lesson Type
      type: object
      title: LessonUpdateRequest
...
    PublicScheduleDayResponse:
      properties:
        date:
          type: string
          format: date
          title: Date
        weekday:
          type: integer
          title: Weekday
        lessons:
          items:
            "$ref": "#/components/schemas/LessonResponse"
          type: array
          title: Lessons
      type: object
      required:
      - date
      - weekday
      title: PublicScheduleDayResponse
...
    PublicScheduleWeekResponse:
      properties:
        week_start:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Week Start
        week_end:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Week End
        week_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week Number
        days:
          items:
            "$ref": "#/components/schemas/PublicScheduleDayResponse"
          type: array
          title: Days
      type: object
      title: PublicScheduleWeekResponse
...
    ScheduleProblemResponse:
      properties:
        severity:
          type: string
          title: Severity
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
        week_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Week Number
        time_slot:
          anyOf:
          - type: integer
          - type: 'null'
          title: Time Slot
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        teacher_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Name
        room_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Room Name
        lesson_ids:
          items:
            type: integer
          type: array
          title: Lesson Ids
      type: object
      required:
      - severity
      - code
      - message
      title: ScheduleProblemResponse
    ScheduleSlotRoomResponse:
      properties:
        room_name:
          type: string
          title: Room Name
        building:
          type: string
          title: Building
        room_is_excluded:
          type: boolean
          title: Room Is Excluded
          default: false
        room_exclusion_reason:
          type: string
          title: Room Exclusion Reason
          default: ''
        lesson:
          anyOf:
          - "$ref": "#/components/schemas/LessonResponse"
          - type: 'null'
      type: object
      required:
      - room_name
      - building
      - lesson
      title: ScheduleSlotRoomResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...